requirements.txt
pysftp
django==1.4.1
django-jenkins
south
vectorformats
psycopg2==2.5
simplejson
openpyxl==1.5.8
xlrd
pysolr
pep8
mox
mock
python-dateutil
django_coverage
django-userena
django-guardian==1.1.1
django-nose
python-memcached
django-memcached
django-tastypie==0.9.15
django-orm-extensions
-e git://github.com/hcarvalhoalves/django-rest-thumbnails.git@6e949d137246e9d3d8180dbc3fa26c512e753b11#egg=django-rest-thumbnails
httplib2
model_mommy==0.8.1
django-waffle
nose-exclude
django-dajaxice
django-dajax
django-bootstrap-toolkit
